feat PA-28: changed file web/password_validation.py
