--- backend/web/password_validation.py
@@ -1,63 +1,0 @@
-import re
-from django.core.exceptions import ValidationError
-from django.utils.translation import ugettext as _
-
-
-class PasswordMinimumLengthValidator:
-    def __init__(self, min_length=8):
-        self.min_length = min_length
-
-    def validate(self, password):
-        if len(password) < self.min_length:
-            raise ValidationError(
-                _("This password must contain at least %(min_length)d characters."),
-                code='password_too_short',
-                params={'min_length': self.min_length},
-            )
-
-    def get_help_text(self):
-        return _(
-            f"Your password must contain at least {self.min_length} characters."
-        )
-
-
-class DigitInPasswordValidator(object):
-    def validate(self, password):
-        if not re.findall(r'\d', password):
-            raise ValidationError(
-                _("The password must contain at least %(min_digits)d digit(s), 0-9."),
-                code='password_no_number',
-            )
-
-    def get_help_text(self):
-        return _(
-            "Your password must contain at least 1 digit, 0-9."
-        )
-
-
-class UppercaseInPasswordValidator(object):
-    def validate(self, password):
-        if not re.findall(r'[A-Z]', password):
-            raise ValidationError(
-                _("The password must contain at least 1 uppercase letter, A-Z."),
-                code='password_no_upper',
-            )
-
-    def get_help_text(self):
-        return _(
-            "Your password must contain at least 1 uppercase letter, A-Z."
-        )
-
-
-class LowercaseInPasswordValidator(object):
-    def validate(self, password):
-        if not re.findall('[a-z]', password):
-            raise ValidationError(
-                _("The password must contain at least 1 lowercase letter, a-z."),
-                code='password_no_lower',
-            )
-
-    def get_help_text(self):
-        return _(
-            "Your password must contain at least 1 lowercase letter, a-z."
-        )
